Fixed saving error, now the model is saved at each step and the old one is canceled

+++ birmingham_cell_tests/src/td3_and_gym_env.py
@@ -16,27 +16,32 @@ import os
 class MyCheckpointCallback(CheckpointCallback):
     def __init__(self, save_freq: int, save_path: str, name_prefix: str = "rl_model"):
         super(MyCheckpointCallback, self).__init__(save_freq=save_freq, save_path=save_path, name_prefix=name_prefix)
-        self.episode_number = 0
-
-    def _on_training_start(self) -> None:
-        # Rimuovi il modello precedente all'inizio dell'addestramento
-        self._remove_previous_models()
 
     def _on_step(self) -> bool:
+        if self.n_calls % self.save_freq == 0:
+            for filename in os.listdir(self.save_path):
+                if filename.startswith(self.name_prefix):
+                    os.remove(os.path.join(self.save_path, filename))
+            model_path = self._checkpoint_path(extension="zip")
+            self.model.save(model_path)
+            if self.verbose >= 2:
+                print(f"Saving model checkpoint to {model_path}")
+
+            if self.save_replay_buffer and hasattr(self.model, "replay_buffer") and self.model.replay_buffer is not None:
+                # If model has a replay buffer, save it too
+                replay_buffer_path = self._checkpoint_path("replay_buffer_", extension="pkl")
+                self.model.save_replay_buffer(replay_buffer_path)  # type: ignore[attr-defined]
+                if self.verbose > 1:
+                    print(f"Saving model replay buffer checkpoint to {replay_buffer_path}")
+
+            if self.save_vecnormalize and self.model.get_vec_normalize_env() is not None:
+                # Save the VecNormalize statistics
+                vec_normalize_path = self._checkpoint_path("vecnormalize_", extension="pkl")
+                self.model.get_vec_normalize_env().save(vec_normalize_path)  # type: ignore[union-attr]
+                if self.verbose >= 2:
+                    print(f"Saving model VecNormalize to {vec_normalize_path}")
+
         return True
-
-    def _on_episode_end(self) -> None:
-        self.episode_number += 1
-        # Rimuovi il modello precedente prima di salvare il nuovo modello
-        self._remove_previous_models()
-        super(MyCheckpointCallback, self)._on_episode_end()
-
-    def _remove_previous_models(self) -> None:
-        # Rimuovi i modelli precedenti
-        for filename in os.listdir(self.save_path):
-            if filename.startswith(self.name_prefix):
-                os.remove(os.path.join(self.save_path, filename))
-
 
 # from gymnasium import NormalizeObservation
 # from stable_baselines3.common.vec_env import VecNormalize
@@ -47,10 +52,10 @@ epoch_number = 500
 max_epoch_steps = 60
 learning_start_steps = 40
 train_freq = 1
-learning_rate = 0.1
+learning_rate = 0.01
 gamma = 0.9
 total_timesteps = max_epoch_steps * epoch_number
-model_save_freq = 100
+model_save_freq = 1
 
 rospack = rospkg.RosPack()
 path = rospack.get_path('birmingham_cell_tests')
@@ -71,8 +76,8 @@ env = gym.make('ConnectionEnv-v0',
                action_type='increment_value', 
                max_episode_steps=max_epoch_steps, 
                data_file_name=data_name,
-               debug_mode=True)
-            #    debug_mode=False)
+            #    debug_mode=True)
+               debug_mode=False)
 
 # env = gym.NormalizeObservation(env)
 
@@ -82,7 +87,7 @@ env = gym.make('ConnectionEnv-v0',
 #    name_prefix=model_name,
 #    keep_only_best=False
 #)
-checkpoint_callback = MyCheckpointCallback(save_freq=max_epoch_steps,
+checkpoint_callback = MyCheckpointCallback(save_freq=model_save_freq,
                                            save_path=models_repo_path + '/', 
                                            name_prefix=model_name)
